shrink the signaling section
The working group has decided that we do not need to tackle specifying
the signaling protocol. Rather than leaving the document with it
partially specified, remove most of the text related to it. Keep some
basic examples of constraints.

This also removes the text which claimed that the document would specify
rqeuirements for the protocol. It does not.

Made a few related editorial changes.