rephrase supported api version section
Everyone was confused by the phrasing that the API needed to provide
evidence it supported the architecture. Since it's a proper component of
the architecture, it doesn't really need to; that is assumed. Rather, it
needs to inform the client of versions/types/etc so that the client can
determine whether *it* supports the API. Rephrase accordingly.