update component diagram to mention optional visit
Whether or not the user visits the User Portal is optional: the user
could choose not to, or the Captive Portal may not provide one. Clarify
that the case presented in the component diagram is the one where the
User Portal is present, and whwere the user chooses to visit it.