Mention state in UE section
In section 4.2, we describe a condition where the session is about to
expire. The UE detects this by consulting some form of storage. Yet, we
never mentioned that it was proper to store any state.

This commit adds a small section outlining that the UE can store state,
with a brief discussion about the tradeoffs in doing so.