add an api requirement that UE alone can request
A reviewer pointed out that we never really said that it was a desired
proprety of the API that it limit access to a UE's state to itself. Add
a requirement to that effect so we're explicit about that goal, and
reference the identifier section that describes that it's a good
property for the identity.