replace Captive Network with Captive Portal
A reviewer pointed out that Captive Portal and Captive Network were used
interchangeably in the document. While we used Captive Network more
frequently, Captive Portal aligns more closely with familiar use of the
term. Further, other documents such as the API draft refer to the system
as a whole as a Captive Portal, which makes this change more aligned
with them.